\name{toydata}
\alias{toydata}
\title{A toy dataset to run examples}
\description{
  A toy dataset to run examples in this package.
}

\details{
  This data includes:  
    a Seurat object `seu`, which is a subset of Xenium breast cancer section;
    a data.frame `pos`, which is the location information of spots;
    a list `pathway_list` contains two pathway of intersecting;
    a list `markers` contains marker genes for each cell types;
    a matrix `imgf` contains histology image features. 
}
